fix(runs): strong of autokeras trials
Sorting for trials seemed broken, because it was sotring based on the
highest value of he metric over the epochs, which is obviously not what
yuou want. So now when sorting from low to  high it looks at the lowest
temporal value and when sorting the other way around, it takes the
highest temporal value.