refactor(pruning): pruning is now on layer base
whilst pruning as applied to the model as a whole in the past, it is nwo
applied on a layer base. This allows us in the future to introduce a
fature to individually prune the layers and helps with models that
contain some kind of unprunable layer, We now can justleave this layer
out and the erest of the model still is pruned.