fix(timing): measure time in first separate callback
measuring time for ther computation is not straight forward. First of
all it needss to be the first thing to do in the epoch or prediction
callback, secondly the energy measurement takes a lot of time, so in
prediction we can not measure enery for every datapoint. Thats why we
introduced the energy measurement frequency to only measure energy at a
certain time. also then do not measure the time as this measurement
would be wrong.